+++ emu/nds/rast/gl/color.go
@@ -10,7 +10,7 @@ var (
 // Color uses range 0..1
 
 type Color struct {
-	R, G, B, A float64
+	R, G, B, A float32
 }
 
 func MakeColorFrom15Bit(r, g, b uint8) Color {
@@ -20,13 +20,13 @@ func MakeColorFrom15Bit(r, g, b uint8) Color {
 	b = (b << 3) | (b >> 2)
 
 	const d = 0xff
-	return Color{float64(r) / d, float64(g) / d, float64(b) / d, 1}
+	return Color{float32(r) / d, float32(g) / d, float32(b) / d, 1}
 }
 
 func (a Color) Add(b Color) Color {
 	return Color{a.R + b.R, a.G + b.G, a.B + b.B, a.A + b.A}
 }
 
-func (a Color) MulScalar(b float64) Color {
+func (a Color) MulScalar(b float32) Color {
 	return Color{a.R * b, a.G * b, a.B * b, a.A * b}
 }
